+++ src/Processor.ts
@@ -25,7 +25,7 @@ import * as _ from "lodash";
 import {
   IDownload,
   IIntent,
-  Invocation,
+  IInvocation,
   IPublishingInformation,
   ISlot,
   ISlotSynonymns,
@@ -34,9 +34,13 @@ import {
 import { getSheetType, IVoxaSheet, SheetTypes } from "./VoxaSheet";
 
 export function sheetLocale(voxaSheet: IVoxaSheet, AVAILABLE_LOCALES: string[]) {
-  let locale = AVAILABLE_LOCALES.find((loc: string) =>
-    _.endsWith(_.toLower(voxaSheet.spreadsheetTitle), _.toLower(loc))
-  );
+  let locale = AVAILABLE_LOCALES.find((loc: string) => {
+    let lastDotIndex = _.lastIndexOf(voxaSheet.spreadsheetTitle, ".");
+    lastDotIndex = lastDotIndex === -1 ? voxaSheet.spreadsheetTitle.length : lastDotIndex;
+
+    const titleWithoutExtension = _.toLower(voxaSheet.spreadsheetTitle.slice(0, lastDotIndex));
+    return _.endsWith(titleWithoutExtension, _.toLower(loc));
+  });
   locale = locale || AVAILABLE_LOCALES[0];
 
   return locale;
@@ -77,12 +81,12 @@ export function invocationProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES:
 
       return acc;
     },
-    [] as Invocation[]
+    [] as IInvocation[]
   );
 }
 
 export function viewsProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES: string[]) {
-  function sanitizeView(text: string) {
+  function sanitizeView(text: string = "") {
     return text
       .replace(/’/g, "'")
       .replace(/’/g, "'")
@@ -105,20 +109,20 @@ export function viewsProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES: stri
         if (_.isEmpty(path)) {
           return acc;
         }
-        const shouldBeArray = [".say", ".reprompt", ".tell", ".ask"].find(suffix =>
+        const shouldBeArray = [".text", ".say", ".reprompt", ".tell", ".ask"].find(suffix =>
           path.includes(suffix)
         );
         const isASuggestionChip = [".dialogflowsuggestions", ".facebooksuggestionchips"].find(
           option => pathLowerCase.includes(option)
         );
 
-        if (shouldBeArray && value) {
-          const temp = (acc as any)[path] || [];
+        if (shouldBeArray && _.isString(value) && !_.isEmpty(value)) {
+          const temp = _.get(acc, path, []) as string[];
           temp.push(sanitizeView(value));
           value = temp;
         }
 
-        if (isASuggestionChip) {
+        if (!_.isEmpty(value) && isASuggestionChip) {
           value = value.split("\n").map((v: string) => v.trim());
         }
 
@@ -166,35 +170,22 @@ export function slotProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES: strin
 }
 
 export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES: string[]) {
-  const voxaSheetsIntent = voxaSheets.filter(voxaSheet =>
-    _.includes([SheetTypes.INTENT], getSheetType(voxaSheet))
+  const voxaSheetsIntent = filterSheets(voxaSheets, [SheetTypes.INTENT]);
+
+  const voxaSheetsUtter = _.reduce(
+    filterSheets(voxaSheets, [SheetTypes.UTTERANCE]),
+    reduceIntent("utterance"),
+    [] as IVoxaSheet[]
   );
-  let voxaSheetsUtter = voxaSheets.filter(voxaSheet => {
-    return _.includes([SheetTypes.UTTERANCE], getSheetType(voxaSheet));
-  });
 
-  voxaSheetsUtter = _.chain(voxaSheetsUtter)
-    .reduce(
-      (acc, utter) => {
-        utter.data = _.chain(utter.data)
-          .reduce((accData: Array<{}>, item: any) => {
-            _.map(item, (value, key) => {
-              accData.push({ intent: key, utterance: value });
-            });
-            return accData;
-          }, [])
-          .groupBy("intent")
-          .value();
-
-        acc.push(utter);
-        return acc;
-      },
-      [] as IVoxaSheet[]
-    )
-    .value();
+  const voxaSheetResponses = _.reduce(
+    filterSheets(voxaSheets, [SheetTypes.RESPONSES]),
+    reduceIntent("response"),
+    [] as IVoxaSheet[]
+  );
 
   const result = _.chain(voxaSheetsIntent)
-    .map(voxaSheetIntent => {
+    .map((voxaSheetIntent: IVoxaSheet) => {
       const locale = sheetLocale(voxaSheetIntent, AVAILABLE_LOCALES);
       let previousIntent: string;
       voxaSheetIntent.data = _.chain(voxaSheetIntent.data)
@@ -203,6 +194,7 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
             "Intent",
             "slotType",
             "slotName",
+            "slotRequired",
             "environment",
             "platformIntent",
             "events",
@@ -210,7 +202,9 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
             "startIntent",
             "signInRequired",
             "endIntent",
-            "platformSlot"
+            "platformSlot",
+            "webhookForSlotFilling",
+            "webhookUsed"
           ]);
           previousIntent = _.isEmpty(info.Intent) ? previousIntent : info.Intent;
           info.Intent = previousIntent;
@@ -221,7 +215,7 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
         .groupBy("Intent")
         .toPairs()
         .reduce(
-          (acc, item) => {
+          (acc: IIntent[], item: any) => {
             const intentName = item[0] as string;
             const head = _.head(item[1]);
             const events = _.chain(head)
@@ -245,25 +239,34 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
               .compact()
               .value() as string[];
 
+            const webhookForSlotFilling = (_.get(head, "webhookForSlotFilling", false) ||
+              _.get(head, "useWebhookForSlotFilling", false)) as boolean;
+            const webhookUsed = _.get(head, "webhookUsed", true) as boolean;
             const canFulfillIntent = _.get(head, "canFulfillIntent", false) as boolean;
             const startIntent = _.get(head, "startIntent", false) as boolean;
             const endIntent = _.get(head, "endIntent", false) as boolean;
 
-            const samples = _(voxaSheetsUtter)
-              .filter({ spreadsheetId: voxaSheetIntent.spreadsheetId })
-              .map(spreadSheet => spreadSheet.data[intentName] || [])
-              .flatten()
-              .map("utterance")
-              .compact()
-              .uniq()
-              .value();
+            const samples = getIntentValueList(
+              voxaSheetsUtter,
+              voxaSheetIntent.spreadsheetId,
+              intentName,
+              "utterance"
+            );
+
+            const responses = getIntentValueList(
+              voxaSheetResponses,
+              voxaSheetIntent.spreadsheetId,
+              intentName,
+              "response"
+            );
 
             const slotsDefinition = _.chain(item[1])
               .filter("slotName")
-              .map(slot => ({
+              .map((slot: any) => ({
                 name: slot.slotName,
                 type: slot.slotType,
-                platform: slot.platformSlot
+                platform: slot.platformSlot,
+                required: slot.slotRequired || false
               }))
               .compact()
               .uniq()
@@ -272,7 +275,10 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
             const intent: IIntent = {
               name: intentName,
               samples,
+              responses,
               slotsDefinition,
+              webhookForSlotFilling,
+              webhookUsed,
               canFulfillIntent,
               startIntent,
               endIntent,
@@ -295,7 +301,6 @@ export function intentUtterProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES
     .flattenDeep()
     .value();
 
-  // console.log('result', result);
   return result as IIntent[];
 }
 
@@ -337,4 +342,43 @@ export function publishingProcessor(voxaSheets: IVoxaSheet[], AVAILABLE_LOCALES:
     },
     [] as IPublishingInformation[]
   );
+}
+
+function filterSheets(voxaSheets: IVoxaSheet[], sheetTypes: string[]): IVoxaSheet[] {
+  return _.filter(voxaSheets, voxaSheet => _.includes(sheetTypes, getSheetType(voxaSheet)));
+}
+
+function reduceIntent(propName: string) {
+  return (acc: any, row: any) => {
+    row.data = _.chain(row.data)
+      .reduce((accData: Array<{}>, item: any) => {
+        _.map(item, (value, key) => {
+          const obj: any = { intent: key };
+          obj[propName] = value;
+          accData.push(obj);
+        });
+        return accData;
+      }, [])
+      .groupBy("intent")
+      .value();
+
+    acc.push(row);
+    return acc;
+  };
+}
+
+function getIntentValueList(
+  voxaSheets: IVoxaSheet[],
+  spreadsheetId: string,
+  intentName: string,
+  key: string
+): string[] {
+  return _(voxaSheets)
+    .filter(sheet => sheet.spreadsheetId === spreadsheetId)
+    .map((spreadSheet: IVoxaSheet) => spreadSheet.data[intentName] || [])
+    .flatten()
+    .map(key)
+    .compact()
+    .uniq()
+    .value();
 }