+++ src/connectors/Excel.ts
@@ -40,7 +40,7 @@ function findLocalFiles(spreadsheet: string): string[] {
 
 function readFileCreateWorkbook(f: string) {
   const workbook = xlsx.parse(f);
-  const spreadsheetTitle = _.last(f.split("/"));
+  const spreadsheetTitle = _.last(f.split("/")) as string;
   const spreadsheetId = f;
   if (!_.isString(spreadsheetTitle)) {
     return undefined;
@@ -50,7 +50,7 @@ function readFileCreateWorkbook(f: string) {
     spreadsheetTitle,
     sheetTitle: book.name,
     type: "none",
-    data: processBookData(book.data)
+    data: processBookData(book.data).filter(row => row.length)
   }));
 }