+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [2.1.2] - 2019-05-08
+
+### Added
+
+- Added more locales to Alexa and Dialogflow
+- Added support for webhookForSlotFilling (dialogflow)
+- Added support for webhookUsed (dialogflow)
+- Added support for intent responses (dialogflow)
+- Added support for required slots
+
+### Fixed
+
+- Fixed views overwriting existing array
+- Fixed error on publishing information overwriting keys
+- Filter out empty rows when processing a local spreadsheet
+- Fixed empty suggestion chip on VUI
+- Fixed overlap on utterance and slots
+- Fixed views have an array path with object
+- Fixed issue with pt-br is not attached to its parent
+
+### Removed
+
+- Removed unused dependencies
+
 ## [2.1.1] - 2019-03-29
 
 ### Fixed
@@ -76,8 +102,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for Alexa, Dialogflow
 - Add support to download other sheet
 
-[unreleased]: https://github.com/VoxaAI/voxa-cli/compare/2.1.1...staging
-[2.1.0]: https://github.com/VoxaAI/voxa-cli/compare/2.1.0...2.1.1
+[unreleased]: https://github.com/VoxaAI/voxa-cli/compare/2.1.2...staging
+[2.1.2]: https://github.com/VoxaAI/voxa-cli/compare/2.1.1...2.1.2
+[2.1.1]: https://github.com/VoxaAI/voxa-cli/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/VoxaAI/voxa-cli/compare/2.0.1...2.1.0
 [2.0.1]: https://github.com/VoxaAI/voxa-cli/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/VoxaAI/voxa-cli/compare/1.0.0-alpha1...2.0.0